Refactor front-end structure
