+++ src/web/.gitignore
@@ -22,4 +22,4 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-src/firebase/fbconfig.js
+src/configs/firebase/fbconfig.js
