+++ src/web/src/constants/index.js
@@ -1,0 +1,1 @@
+export { popUpFormNames } from './pop-up-forms-name';
