+++ src/web/src/constants/pop-up-forms-name.js
@@ -1,0 +1,3 @@
+export const popUpFormNames = {
+  JOIN_EXPENSE: 'JOIN_EXPENSE_FORM'
+};
