--- src/web/src/contexts/SWContext.js
@@ -1,6 +1,0 @@
-import { createContext } from 'react';
-
-export const SWContext = createContext({
-  loggedInUser: {},
-  setLoggedInUser: () => {}
-});
