+++ src/web/src/features/expense/components/ExpenseTableForms/index.js
@@ -1,0 +1,4 @@
+export * from './AddReceiptImgForm';
+export * from './SplitForm';
+export * from './ReceiptForm';
+export * from './ReceiptImgForm';
