+++ src/web/src/features/expense/components/index.js
@@ -1,0 +1,2 @@
+export * from './Expense';
+export * from './ExpenseList';
