+++ src/web/src/layouts/RootLayout.js
@@ -1,0 +1,18 @@
+import React, { useContext } from 'react';
+import { PopUpForm } from '../features';
+
+// workspace imports
+
+import { AppContext } from '../providers';
+
+const RootLayout = ({ children }) => {
+  const { popUpForm } = useContext(AppContext);
+  return (
+    <>
+      {popUpForm ? <PopUpForm formName={popUpForm} /> : null}
+      {children}
+    </>
+  );
+};
+
+export { RootLayout };
