+++ src/web/src/layouts/index.js
@@ -1,0 +1,1 @@
+export { RootLayout } from './RootLayout';
