+++ src/web/src/pages/index.js
@@ -1,0 +1,1 @@
+export { Router } from './router';
