+++ src/web/src/providers/index.js
@@ -1,0 +1,1 @@
+export * from './AppProvider';
